requirements.txt
torch==2.5.1
torchao==0.8.0
mne==1.9.0
pandas==2.2.3
numpy==2.2.5
scikit-learn==1.5.2
scipy==1.14.1
matplotlib==3.9.2
tqdm==4.67.0
python-dotenv==1.0.1